[REFACTOR] UserInfo 관련 dto 수정 및 파일 위치 이동
