+++ src/main/java/com/triprecord/triprecord/user/dto/request/UserCreateRequest.java
@@ -1,4 +1,4 @@
-package com.triprecord.triprecord.user.controller.dto.request;
+package com.triprecord.triprecord.user.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotNull;
